Move UTM parameter tracking logic into stencil component
